set old branch as deprecated
