--- src/main/java/usecases/dataStorage/account/AccountDsGateway.java
@@ -1,6 +1,0 @@
-package usecases.dataStorage.account;
-
-public interface AccountDsGateway {
-    boolean existsAccount(AccountDsRequestModel requestModel);
-    AccountDsModel loadAccount(AccountDsRequestModel requestModel) throws IllegalArgumentException;
-}
