--- src/main/java/usecases/dataStorage/account/AccountDsRequestModel.java
@@ -1,13 +1,0 @@
-package usecases.dataStorage.account;
-
-public class AccountDsRequestModel {
-    private String username;
-
-    public AccountDsRequestModel(String username) {
-        this.username = username;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-}
