--- src/main/java/usecases/dataStorage/assessment/AssessmentDsGateway.java
@@ -1,6 +1,0 @@
-package usecases.dataStorage.assessment;
-
-public interface AssessmentDsGateway {
-    boolean existsAssessment(AssessmentDsRequestModel requestModel);
-    AssessmentDsModel loadAssessment(AssessmentDsRequestModel requestModel) throws IllegalArgumentException;
-}
