--- src/main/java/usecases/dataStorage/assessment/AssessmentDsRequestModel.java
@@ -1,16 +1,0 @@
-package usecases.dataStorage.assessment;
-
-import usecases.dataStorage.course.CourseDsRequestModel;
-
-public class AssessmentDsRequestModel extends CourseDsRequestModel {
-    private String assessmentTitle;
-
-    public AssessmentDsRequestModel(String username, String courseCode, String assessmentTitle) {
-        super(username, courseCode);
-        this.assessmentTitle = assessmentTitle;
-    }
-
-    public String getAssessmentTitle() {
-        return assessmentTitle;
-    }
-}
