--- src/main/java/usecases/dataStorage/course/CourseDsGateway.java
@@ -1,6 +1,0 @@
-package usecases.dataStorage.course;
-
-public interface CourseDsGateway {
-    boolean existsCourse(CourseDsRequestModel requestModel);
-    CourseDsModel loadCourse(CourseDsRequestModel requestModel) throws IllegalArgumentException;
-}
