--- src/main/java/usecases/dataStorage/instance/InstanceDsGateway.java
@@ -1,7 +1,0 @@
-package usecases.dataStorage.instance;
-
-public interface InstanceDsGateway {
-    boolean existsInstance(InstanceDsRequestModel requestModel);
-    InstanceDsModel loadInstance(InstanceDsRequestModel requestModel) throws IllegalArgumentException;
-    void saveInstanceMark(InstanceDsRequestModel requestModel, double mark) throws IllegalArgumentException;
-}
