--- src/main/java/usecases/dataStorage/instance/InstanceDsRequestModel.java
@@ -1,16 +1,0 @@
-package usecases.dataStorage.instance;
-
-import usecases.dataStorage.assessment.AssessmentDsRequestModel;
-
-public class InstanceDsRequestModel extends AssessmentDsRequestModel {
-    private int instanceIndex;
-
-    public InstanceDsRequestModel(String username, String courseCode, String assessmentTitle, int instanceIndex) {
-        super(username, courseCode, assessmentTitle);
-        this.instanceIndex = instanceIndex;
-    }
-
-    public int getInstanceIndex() {
-        return instanceIndex;
-    }
-}
