--- src/main/java/usecases/dataStorage/weightScheme/OrderedWeightDsModel.java
@@ -1,13 +1,0 @@
-package usecases.dataStorage.weightScheme;
-
-public class OrderedWeightDsModel extends WeightSchemeDsModel{
-    private WeightDsModel[] orderedWeightModels;
-
-    public OrderedWeightDsModel(WeightDsModel[] orderedWeightModels) {
-        this.orderedWeightModels = orderedWeightModels;
-    }
-
-    public WeightDsModel[] getOrderedWeightModels() {
-        return orderedWeightModels;
-    }
-}
