--- src/main/java/usecases/dataStorage/weightScheme/WeightDsModel.java
@@ -1,21 +1,0 @@
-package usecases.dataStorage.weightScheme;
-
-import java.io.Serializable;
-
-public class WeightDsModel implements Serializable {
-    private int numberOfInstances;
-    private double weightOfEachInstance;
-
-    public WeightDsModel(int numberOfInstances, double weightOfEachInstance) {
-        this.numberOfInstances = numberOfInstances;
-        this.weightOfEachInstance = weightOfEachInstance;
-    }
-
-    public int getNumberOfInstances() {
-        return numberOfInstances;
-    }
-
-    public double getWeightOfEachInstance() {
-        return weightOfEachInstance;
-    }
-}
