--- src/main/java/usecases/dataStorage/weightScheme/WeightSchemeDsModel.java
@@ -1,5 +1,0 @@
-package usecases.dataStorage.weightScheme;
-
-import java.io.Serializable;
-
-public class WeightSchemeDsModel implements Serializable { }
