+++ src/main/java/usecases/dataStorage_DEPRECATED/account/AccountDsGateway.java
@@ -1,0 +1,6 @@
+package usecases.dataStorage_DEPRECATED.account;
+
+public interface AccountDsGateway {
+    boolean existsAccount(AccountDsRequestModel requestModel);
+    AccountDsModel loadAccount(AccountDsRequestModel requestModel) throws IllegalArgumentException;
+}
