+++ src/main/java/usecases/dataStorage_DEPRECATED/assessment/AssessmentDsGateway.java
@@ -1,0 +1,6 @@
+package usecases.dataStorage_DEPRECATED.assessment;
+
+public interface AssessmentDsGateway {
+    boolean existsAssessment(AssessmentDsRequestModel requestModel);
+    AssessmentDsModel loadAssessment(AssessmentDsRequestModel requestModel) throws IllegalArgumentException;
+}
