+++ src/main/java/usecases/dataStorage_DEPRECATED/assessment/AssessmentDsRequestModel.java
@@ -1,0 +1,16 @@
+package usecases.dataStorage_DEPRECATED.assessment;
+
+import usecases.dataStorage_DEPRECATED.course.CourseDsRequestModel;
+
+public class AssessmentDsRequestModel extends CourseDsRequestModel {
+    private String assessmentTitle;
+
+    public AssessmentDsRequestModel(String username, String courseCode, String assessmentTitle) {
+        super(username, courseCode);
+        this.assessmentTitle = assessmentTitle;
+    }
+
+    public String getAssessmentTitle() {
+        return assessmentTitle;
+    }
+}
