+++ src/main/java/usecases/dataStorage_DEPRECATED/course/CourseDsGateway.java
@@ -1,0 +1,6 @@
+package usecases.dataStorage_DEPRECATED.course;
+
+public interface CourseDsGateway {
+    boolean existsCourse(CourseDsRequestModel requestModel);
+    CourseDsModel loadCourse(CourseDsRequestModel requestModel) throws IllegalArgumentException;
+}
