+++ src/main/java/usecases/dataStorage_DEPRECATED/instance/InstanceDsGateway.java
@@ -1,0 +1,7 @@
+package usecases.dataStorage_DEPRECATED.instance;
+
+public interface InstanceDsGateway {
+    boolean existsInstance(InstanceDsRequestModel requestModel);
+    InstanceDsModel loadInstance(InstanceDsRequestModel requestModel) throws IllegalArgumentException;
+    void saveInstanceMark(InstanceDsRequestModel requestModel, double mark) throws IllegalArgumentException;
+}
