+++ src/main/java/usecases/dataStorage_DEPRECATED/weightScheme/OrderedWeightDsModel.java
@@ -1,0 +1,13 @@
+package usecases.dataStorage_DEPRECATED.weightScheme;
+
+public class OrderedWeightDsModel extends WeightSchemeDsModel{
+    private WeightDsModel[] orderedWeightModels;
+
+    public OrderedWeightDsModel(WeightDsModel[] orderedWeightModels) {
+        this.orderedWeightModels = orderedWeightModels;
+    }
+
+    public WeightDsModel[] getOrderedWeightModels() {
+        return orderedWeightModels;
+    }
+}
