+++ src/main/java/usecases/dataStorage_DEPRECATED/weightScheme/WeightSchemeDsModel.java
@@ -1,0 +1,5 @@
+package usecases.dataStorage_DEPRECATED.weightScheme;
+
+import java.io.Serializable;
+
+public class WeightSchemeDsModel implements Serializable { }
