--- src/main/java/usecases/markOperations/addMark/AddMarkInputBoundary.java
@@ -1,5 +1,0 @@
-package usecases.markOperations.addMark;
-
-public interface AddMarkInputBoundary {
-    AddMarkResponseModel execute(AddMarkRequestModel requestModel);
-}
