--- src/main/java/usecases/markOperations/addMark/AddMarkPresenter.java
@@ -1,6 +1,0 @@
-package usecases.markOperations.addMark;
-
-public interface AddMarkPresenter {
-    AddMarkResponseModel prepareSuccessView(AddMarkResponseModel responseModel);
-    AddMarkResponseModel prepareFailureView(String errorMessage);
-}
