--- src/main/java/usecases/markOperations/addMark/AddMarkResponseModel.java
@@ -1,13 +1,0 @@
-package usecases.markOperations.addMark;
-
-public class AddMarkResponseModel {
-    private double newHypotheticalGrade;
-
-    public AddMarkResponseModel(double newHypotheticalGrade) {
-        this.newHypotheticalGrade = newHypotheticalGrade;
-    }
-
-    public double getNewHypotheticalGrade() {
-        return newHypotheticalGrade;
-    }
-}
