+++ src/main/java/usecases/markOperations_DEPRECATED/addMark/AddMarkDsGateway.java
@@ -1,0 +1,7 @@
+package usecases.markOperations_DEPRECATED.addMark;
+
+import usecases.dataStorage_DEPRECATED.account.AccountDsGateway;
+import usecases.dataStorage_DEPRECATED.instance.InstanceDsGateway;
+
+public interface AddMarkDsGateway extends AccountDsGateway, InstanceDsGateway {
+}
