+++ src/main/java/usecases/markOperations_DEPRECATED/addMark/AddMarkInputBoundary.java
@@ -1,0 +1,5 @@
+package usecases.markOperations_DEPRECATED.addMark;
+
+public interface AddMarkInputBoundary {
+    AddMarkResponseModel execute(AddMarkRequestModel requestModel);
+}
