+++ src/main/java/usecases/markOperations_DEPRECATED/addMark/AddMarkPresenter.java
@@ -1,0 +1,6 @@
+package usecases.markOperations_DEPRECATED.addMark;
+
+public interface AddMarkPresenter {
+    AddMarkResponseModel prepareSuccessView(AddMarkResponseModel responseModel);
+    AddMarkResponseModel prepareFailureView(String errorMessage);
+}
