+++ src/main/java/usecases/markOperations_DEPRECATED/addMark/AddMarkResponseModel.java
@@ -1,0 +1,13 @@
+package usecases.markOperations_DEPRECATED.addMark;
+
+public class AddMarkResponseModel {
+    private double newHypotheticalGrade;
+
+    public AddMarkResponseModel(double newHypotheticalGrade) {
+        this.newHypotheticalGrade = newHypotheticalGrade;
+    }
+
+    public double getNewHypotheticalGrade() {
+        return newHypotheticalGrade;
+    }
+}
